MATLAB Code
This MATLAB script offers a well-rounded approach for simulating and analyzing the ECG signal in a way that would realistically replicate it from an actual heart. An ECG signal is created with different heart rates and fed with realistic noise and baseline wandering meant to mimic those caused by the body. It is then subjected to bandpass Butterworth filtering to retrieve an unadulterated version before subjecting it to R-peak detection, which is the basis for deriving heart rate values. The analysis then goes into ways of visualizing noisy and filtered signals, as well as their combined displays in reference to heart rate trends. Equipped with customizable parameters, this is a device that will allow any individual to study ECG signal processing in detail to take them deeper into possible conclusions on the cardiac rhythm as supportive biomedical research as well as educational applications.